refactor: apply TigerStyle safety and naming rules to sql module
Critical:
- Split executeScan (1069L) into executeScanInner dispatch + 9 per-variant
  helpers (executeScanBase/Ann/Filter/Project/Limit/Sort/Window/HashJoin/
  HashAgg); add MAX_PLAN_DEPTH=64 depth parameter bounding the recursion
- Split planSelect (330L) into planSelectJoins/JoinUsing/Where/GroupBy/
  Window/OrderLimit/Project helpers, each under 70 lines
- Split evalBuiltin (193L) into evalBuiltinPredicate/String/Math
- Split evalBinary (170L) into evalBinaryArith/Bitwise/Json/JsonAccess

Major:
- Add StorageReadError to SqlExecError; change all iter.next()/storage.get()
  catch null to catch return error.StorageReadError — silent truncation of
  scan results on I/O errors is now a propagated failure
- cdc.dispatch catch {} → store error in error_detail; capture_before_images
  catch null → log detail and skip dispatch rather than silently continue
- result_ring_size: usize → u32; add comptime assert for power-of-two
  invariant (ring index math requires this)
- LIMIT/OFFSET @intCast: add negative-value guard (return TypeMismatch)

Moderate:
- parser.zig: add MAX_PARSE_ITEMS=4096 constant and assert at the head of
  all 29 while(true) collection loops
- params_codec: pos: *usize → *u32 in decodeParamValue/decodeParamValueByTag
- Add const assert imports to eval_expr.zig, plan.zig, parser.zig
- Add assert(name.len > 0) to evalBuiltin; assert operator sets in
  evalBinaryArith/Bitwise for paired positive/negative space coverage

Minor:
- planValueToColumnValue catch null: add comments documenting SQL NULL
  semantics intent (type mismatch → SQL NULL is intentional)

+++ src/sql/params_codec.zig
@@ -92,7 +92,7 @@ pub fn decodeParams(data: []const u8, types: []const ast.SqlType, alloc: std.mem
     if (count == 0) return &.{};
     if (types.len > 0 and count != types.len) return error.TypeMismatch;
     const values = try alloc.alloc(ColumnValue, count);
-    var pos: usize = 2;
+    var pos: u32 = 2;
     for (0..count) |i| {
         if (pos >= data.len) return error.TypeMismatch;
         const tag_byte = data[pos];
@@ -106,7 +106,7 @@ pub fn decodeParams(data: []const u8, types: []const ast.SqlType, alloc: std.mem
     return values;
 }
 
-fn decodeParamValue(data: []const u8, pos: *usize, typ: ast.SqlType, alloc: std.mem.Allocator) !ColumnValue {
+fn decodeParamValue(data: []const u8, pos: *u32, typ: ast.SqlType, alloc: std.mem.Allocator) !ColumnValue {
     return switch (typ) {
         .bool => blk: {
             const b = data[pos.*];
@@ -181,7 +181,7 @@ fn decodeParamValue(data: []const u8, pos: *usize, typ: ast.SqlType, alloc: std.
     };
 }
 
-fn decodeParamValueByTag(data: []const u8, pos: *usize, tag: u8, alloc: std.mem.Allocator) !ColumnValue {
+fn decodeParamValueByTag(data: []const u8, pos: *u32, tag: u8, alloc: std.mem.Allocator) !ColumnValue {
     const col_type: ColumnType = switch (tag) {
         0 => .bool_t,
         1 => .int8,
